+++ lib/src/analyzers/lint_analyzer/metrics/metrics_list/halstead_volume/halstead_volume_metric.dart
@@ -43,7 +43,7 @@ class HalsteadVolumeMetric extends FunctionMetric<double> {
 
   @override
   MetricComputationResult<double> computeImplementation(
-    Declaration node,
+    AstNode node,
     Iterable<ScopedClassDeclaration> classDeclarations,
     Iterable<ScopedFunctionDeclaration> functionDeclarations,
     InternalResolvedUnitResult source,