+++ lib/src/analyzers/lint_analyzer/metrics/metrics_list/number_of_parameters_metric.dart
@@ -36,7 +36,7 @@ class NumberOfParametersMetric extends FunctionMetric<int> {
 
   @override
   bool supports(
-    Declaration node,
+    AstNode node,
     Iterable<ScopedClassDeclaration> classDeclarations,
     Iterable<ScopedFunctionDeclaration> functionDeclarations,
     InternalResolvedUnitResult source,
@@ -61,7 +61,7 @@ class NumberOfParametersMetric extends FunctionMetric<int> {
 
   @override
   MetricComputationResult<int> computeImplementation(
-    Declaration node,
+    AstNode node,
     Iterable<ScopedClassDeclaration> classDeclarations,
     Iterable<ScopedFunctionDeclaration> functionDeclarations,
     InternalResolvedUnitResult source,