+++ lib/src/analyzers/lint_analyzer/metrics/metrics_list/source_lines_of_code/source_lines_of_code_metric.dart
@@ -42,7 +42,7 @@ class SourceLinesOfCodeMetric extends FunctionMetric<int> {
 
   @override
   MetricComputationResult<int> computeImplementation(
-    Declaration node,
+    AstNode node,
     Iterable<ScopedClassDeclaration> classDeclarations,
     Iterable<ScopedFunctionDeclaration> functionDeclarations,
     InternalResolvedUnitResult source,
@@ -79,7 +79,7 @@ class SourceLinesOfCodeMetric extends FunctionMetric<int> {
   ) =>
       linesWithCode.map((lineIndex) {
         final lineStartLocation = SourceLocation(
-          source.lineInfo.getOffsetOfLine(lineIndex),
+          source.lineInfo.getOffsetOfLine(lineIndex - 1),
           sourceUrl: source.path,
           line: lineIndex,
           column: 0,