+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_first/visitor.dart
@@ -9,7 +9,7 @@ class _Visitor extends RecursiveAstVisitor<void> {
   void visitMethodInvocation(MethodInvocation node) {
     super.visitMethodInvocation(node);
 
-    if (_isIterableOrSubclass(node.realTarget?.staticType) &&
+    if (isIterableOrSubclass(node.realTarget?.staticType) &&
         node.methodName.name == 'elementAt') {
       final arg = node.argumentList.arguments.first;
 
@@ -23,7 +23,7 @@ class _Visitor extends RecursiveAstVisitor<void> {
   void visitIndexExpression(IndexExpression node) {
     super.visitIndexExpression(node);
 
-    if (_isListOrSubclass(node.realTarget.staticType)) {
+    if (isListOrSubclass(node.realTarget.staticType)) {
       final index = node.index;
 
       if (index is IntegerLiteral && index.value == 0) {
@@ -31,20 +31,4 @@ class _Visitor extends RecursiveAstVisitor<void> {
       }
     }
   }
-
-  bool _isIterableOrSubclass(DartType? type) =>
-      _isIterable(type) || _isSubclassOfIterable(type);
-
-  bool _isIterable(DartType? type) => type?.isDartCoreIterable ?? false;
-
-  bool _isSubclassOfIterable(DartType? type) =>
-      type is InterfaceType && type.allSupertypes.any(_isIterable);
-
-  bool _isListOrSubclass(DartType? type) =>
-      _isList(type) || _isSubclassOfList(type);
-
-  bool _isList(DartType? type) => type?.isDartCoreList ?? false;
-
-  bool _isSubclassOfList(DartType? type) =>
-      type is InterfaceType && type.allSupertypes.any(_isList);
 }